requirements.txt
pytest~=7.3.1
python-dotenv~=1.0.0
azure-storage-blob
pyyaml
Jinja2